% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adaline.R
\name{adaline.new}
\alias{adaline.new}
\title{Create Adaline neural network}
\usage{
adaline.new(nbrInputs, nbrOutputs, use.trace = FALSE)
}
\arguments{
\item{nbrInputs}{Number of inputs.}

\item{nbrOutputs}{Number of outputs}

\item{use.trace}{Use eligibility trace or not.
Only for reinforcement learning.}
}
\value{
Object of class Adaline (and inherited classes)
}
\description{
Create Adaline neural network
}
\examples{
# Create Adaline with two inputs, 3 outputs, assign weights manually.
a <- adaline.new(2, 3)
weights <- matrix(c(0, 0, 1, 1, 2, 2), nrow=3, ncol=2, byrow = TRUE)
a$set.weights(weights)
inputs <- c(1, 1.5)
a$eval(inputs)

# Plot what it looks like for output 2.
x <- y <-seq(0,1,0.05)
m <- expand.grid(x,y)
persp(x, y, matrix(a$eval(m)[,2], nrow=length(x)), ticktype="detailed")

# Small training test
inputs <- c(1,1.5)
targets <- c(1, 2, 3)
a$eval(inputs) # Have to call eval() before calling train()
a$train(targets)
a$eval(inputs) # Should have changed a little
persp(x, y, matrix(a$eval(m)[,2], nrow=length(x)), ticktype="detailed")

# Slightly more training
inputs <- matrix(c(0,0,1,1), nrow=2, byrow=TRUE)
a$eval(inputs)
targets <- matrix(c(0,0,0,1,2,3), nrow=2, byrow=TRUE)
for ( i in 1:100 ) {
  a$eval(inputs[1,])
  a$train(targets[1,])
  a$eval(inputs[2,])
  a$train(targets[2,])
}
a$eval(inputs)
persp(x, y, matrix(a$eval(m)[,2], nrow=length(x)), ticktype="detailed")
}
