% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{affine.transformation.new}
\alias{affine.transformation.new}
\title{Create affine transformation "object" with transformation matrices "A" and "b".}
\usage{
affine.transformation.new(A = NULL, b = NULL)
}
\arguments{
\item{A}{blabla}

\item{b}{blabla}
}
\value{
Object of class AffineTransformation.
}
\description{
If "A" or "b" are NULL, then they need to be initialised using some other method.
}
