% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{boltzmann.new}
\alias{boltzmann.new}
\title{"Boltzmann" softmax policy}
\usage{
boltzmann.new(temperature = 1)
}
\arguments{
\item{temperature}{blabla}
}
\value{
Object of class BoltzmannPolicy
}
\description{
"Boltzmann" softmax policy
}
