% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{classifying.function.approximator.new}
\alias{classifying.function.approximator.new}
\title{Create classifying FunctionApproximator}
\usage{
classifying.function.approximator.new(function.approximator, classifier = NULL)
}
\arguments{
\item{function.approximator}{blabla}

\item{classifier}{blabla}
}
\value{
Object of class FunctionApproximator
}
\description{
Create classifying FunctionApproximator
}
