% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interfaces.R
\name{controller.new}
\alias{controller.new}
\title{Create "Controller" object}
\usage{
controller.new()
}
\value{
An object of class Controller
}
\description{
"Interface class" for all kinds of "controllers", i.e. objects
that take an input value vector and deside on a control action to take
based on the input values.
Public methods are:
\itemize{
\item reset: resets controller for new episode
\item get.state: returns current state/input value vector
\item set.state: move into new state, decide on action(s) and return "output"
values. Output values may be discrete or continuous.
\item get.actions: returns last action(s)
\item has.discrete.actions: TRUE if actions are discrete values, false if
continuous (in that case the action value(s) is typically directly
the control value).
}
}
