% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{create.input.matrix}
\alias{create.input.matrix}
\title{Create input matrix with all permutations for one or more inputs.}
\usage{
create.input.matrix(indices, mins, maxs, steps, default.inputs = 0)
}
\arguments{
\item{indices}{blabla}

\item{mins}{blabla}

\item{maxs}{blabla}

\item{steps}{blabla}

\item{default.inputs}{blabla}
}
\description{
Create an "input matrix" for e.g. a neural network or similar that
contains all combinations of values from "mins" to "maxs" with
steps "steps". "indices" indicates the column indices in the resulting
matrix where the values should be inserted. "default.inputs" gives
the values for all other inputs, which thus have constant values.
The number of columns of the returned matrix is the bigger of the maximum
value of "indices" or the length of "default.inputs". The number of
rows depends on the number of possible input value combinations.
}
\examples{
create.input.matrix(c(2,4),c(0,10),c(1,20),c(0.2,2), 0)
}
