% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{create.permutation.matrix}
\alias{create.permutation.matrix}
\title{Create permutation matrix from all vectors in the given list}
\usage{
create.permutation.matrix(list.of.value.vectors)
}
\arguments{
\item{list.of.value.vectors}{blabla}
}
\value{
Matrix that has as many columns as there are vectors in the
"list.of.value.vectors" list.
}
\description{
This function takes a list of vectors as parameter. Every vector
contains a list of values. The function then creates a matrix that
has as many rows as necessary to contain all the permutations of the
values in the vectors.
}
\examples{
l<-list(c(1,2,3),c(4,5,6)); create.permutation.matrix(l)
}
