% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{create.z.for.persp}
\alias{create.z.for.persp}
\title{Create z matrix for 3D plot, e.g. with "persp".}
\usage{
create.z.for.persp(x, y, f)
}
\arguments{
\item{x}{blabla}

\item{y}{blabla}

\item{f}{blabla}
}
\value{
z matrix with number of rows equal to length of x vector,
number of columns length of y vector
}
\description{
Create z matrix for 3D plot, e.g. with "persp".
}
\examples{
ws<-weighted.sum.new(matrix(c(0.2,0.8),ncol=1))
x <- seq(0, 1, 0.1); y <- seq(0, 1, 0.1)
z <- create.z.for.persp(x, y, ws$eval)
persp(x,y,z)
}
