% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{discretizer.new}
\alias{discretizer.new}
\title{Generic discretizer "class".}
\usage{
discretizer.new(minval, maxval, nintervals, impose.limits = FALSE)
}
\arguments{
\item{minval}{blabla}

\item{maxval}{blabla}

\item{nintervals}{blabla}

\item{impose.limits}{blabla}
}
\value{
Discretizer object.
}
\description{
Returns interval indices
from 1 to "nintervals". "impose.limits" indicates if min/max values
should be used for limiting first/last interval. In that case, the
returned class is zero if the value doesn't fit any interval.
}
