% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{e.greedy.new}
\alias{e.greedy.new}
\title{Epsilon-greedy policy object}
\usage{
e.greedy.new(eps)
}
\arguments{
\item{eps}{blabla}
}
\value{
Object of class EpsilonGreedyPolicy.
}
\description{
Epsilon-greedy policy object
}
