% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EligibilityTrace.R
\name{eligibility.trace.new}
\alias{eligibility.trace.new}
\title{Eligibility trace for the weights of one neural layer.}
\usage{
eligibility.trace.new(ninputs, noutputs, lambda.decay = 0)
}
\arguments{
\item{ninputs}{blabla}

\item{noutputs}{blabla}

\item{lambda.decay}{blabla}
}
\value{
Eligibility trace object
}
\description{
Eligibility trace for the weights of one neural layer, in principle
one weight matrix. Can also be used for lookup-table representations.
}
