% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBF.R
\name{find.best.inka.formula}
\alias{find.best.inka.formula}
\title{Train with formula / data call}
\usage{
find.best.inka.formula(formula, data, ...)
}
\arguments{
\item{formula}{blabla}

\item{data}{blabla}

\item{...}{blabla}
}
\description{
If dependent variable is of type "factor", then it is automatically
one-hot encoded.
}
