% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBF.R
\name{find.best.inka}
\alias{find.best.inka}
\title{Create "n" RBF nets with given parameters and return the one that
performs the best.}
\usage{
find.best.inka(
  n = 1,
  train.inputs,
  train.outputs,
  max.iter = 1,
  inv.whole.set.at.end = F,
  classification.error.limit = NULL,
  rmse.limit = NULL,
  activation.function = squared.distance.activation,
  output.function = imqe.output.function,
  nrbf = T,
  use.bias = F,
  spread = 0.1,
  c = 0.01,
  test.inputs = NULL,
  test.outputs = NULL
)
}
\arguments{
\item{n}{blabla}

\item{train.inputs}{blabla}

\item{train.outputs}{blabla}

\item{max.iter}{blabla}

\item{inv.whole.set.at.end}{blabla}

\item{classification.error.limit}{Limit of misclassified instances to reach
before ending the training. Either "classification.error.limit" or
"rmse.limit" MUST be provided!}

\item{rmse.limit}{RMSE limit to reach before ending the training.
Either "classification.error.limit" or "rmse.limit" MUST be provided!}

\item{activation.function}{blabla}

\item{output.function}{blabla}

\item{nrbf}{blabla}

\item{use.bias}{blabla}

\item{spread}{blabla}

\item{c}{blabla}

\item{test.inputs}{blabla}

\item{test.outputs}{blabla}
}
\description{
Create "n" RBF nets with given parameters and return the one that
performs the best.
}
