% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interfaces.R
\name{function.approximator.new}
\alias{function.approximator.new}
\title{Create "FunctionApproximator" object}
\usage{
function.approximator.new()
}
\value{
An object of class FunctionApproximator
}
\description{
Create and return FunctionApproximator object. This class is intended to
work as an object-oriented "interface", i.e. as a template for what methods
all objects of class FunctionApproximator have to implement.
}
