% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralLayer.R
\name{gaussian.output.function}
\alias{gaussian.output.function}
\title{Gaussian output function.}
\usage{
gaussian.output.function(activation, spread = 1)
}
\arguments{
\item{activation}{blabla}

\item{spread}{blabla}
}
\value{
Gaussian output value.
}
\description{
Spread parameter may be a single value or a vector, i.e.
same spread for all or individual spread parameters.
The activation value should normally be the squared distance.
}
