% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralLayer.R
\name{identity.output.function}
\alias{identity.output.function}
\title{Output is same as activation.}
\usage{
identity.output.function(activation)
}
\arguments{
\item{activation}{blabla}
}
\value{
"activation" value as such.
}
\description{
For neurons who only do a weighted sum.
}
