% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralLayer.R
\name{imqe.output.function}
\alias{imqe.output.function}
\title{Inverse MultiQuadric Equations function.}
\usage{
imqe.output.function(activation, spread = 1)
}
\arguments{
\item{activation}{blabla}

\item{spread}{blabla}
}
\value{
IMQE value.
}
\description{
Spread parameter may be a single value or a vector, i.e. same spread
for all or individual spread parameters. The activation value should
normally be the squared distance.
}
