% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ms.error}
\alias{ms.error}
\title{Mean squared error between two sets of numbers.}
\usage{
ms.error(actual, estimated)
}
\arguments{
\item{actual}{blabla}

\item{estimated}{blabla}
}
\value{
MSE
}
\description{
Mean squared error between two sets of numbers.
}
