% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{nonlineardssmodel.two.inputs.new}
\alias{nonlineardssmodel.two.inputs.new}
\title{Nonlinear "decision support" model for two inputs.}
\usage{
nonlineardssmodel.two.inputs.new()
}
\description{
Nonlinear "decision support" model for two inputs.
}
