% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{normalize.to.one}
\alias{normalize.to.one}
\title{Normalize values in matrix so that their sum is one (row-wise).}
\usage{
normalize.to.one(vector)
}
\arguments{
\item{vector}{blabla}
}
\value{
Normalised matrix.
}
\description{
Normalize values in matrix so that their sum is one (row-wise).
}
