% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBF.R
\name{predict.RBF}
\alias{predict.RBF}
\title{Caret-like "predict" method with formula.}
\usage{
\method{predict}{RBF}(object, newdata, ...)
}
\arguments{
\item{object}{blabla}

\item{newdata}{blabla}

\item{...}{Same as for generic "predict".}
}
\description{
However, here the default is to return the real-valued output values
also for classification tasks, rather than doing one-hot coding.
}
