% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{random.of.max}
\alias{random.of.max}
\title{Index of maximum value}
\usage{
random.of.max(values)
}
\arguments{
\item{values}{blabla}
}
\value{
Index
}
\description{
Return index of maximum value. If there are more than one, then choose
one randomly.
}
