% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBFclassifier.R
\name{rbf.classifier.new}
\alias{rbf.classifier.new}
\title{Create new RBF classifier.}
\usage{
rbf.classifier.new(
  nbrInputs,
  nbrOutputs,
  activation.function = squared.distance.activation,
  output.function = imqe.output.function
)
}
\arguments{
\item{nbrInputs}{How many inputs.}

\item{nbrOutputs}{How many outputs (RBF "neurons").}

\item{activation.function}{Activation function, typically something like
squared distance.}

\item{output.function}{Output function, typically Gaussian or IMQE.}
}
\value{
Object of class RBFclassifier.
}
\description{
This is typically the hidden layer of an RBF network.
}
\examples{
# Create an empty rnf.classifier and set it up manually with four neurons.
rbfc <- rbf.classifier.new(nbrInputs=2, nbrOutputs=0)
rbfc$init.centroids.grid(c(-1,-1), c(1,1), c(2,2))

# Plot in 3D what the outputs look like.
x <- y <- seq(-1.5,1.5,0.1)
inps <- expand.grid(x,y)
z <- rbfc$eval(inps)
par(mfcol=c(2,2))
for ( i in 1:4 ) {
  zplot <- matrix(z[,i], nrow=length(x), ncol=length(y))
  persp(x,y,zplot,theta=135,phi=45,ticktype="detailed")
}
par(mfcol=c(1,1))
}
