% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbf.restore.R
\name{rbf.restore}
\alias{rbf.restore}
\title{Re-create \code{RBF} object from \code{inka} object.}
\usage{
rbf.restore(
  inka,
  activation.function = squared.distance.activation,
  output.function = imqe.output.function
)
}
\arguments{
\item{inka}{blabla}

\item{activation.function}{blabla}

\item{output.function}{blabla}
}
\value{
\code{RBF} object
}
\description{
Re-create \code{RBF} object from \code{inka} object.
}
\examples{
rbf <- train.inka.formula(Species~., data=iris, spread=0.1, max.iter=20,
classification.error.limit=0)
predict(rbf, newdata=iris[100,])
inka <- rbf$as.inka()
newrbf <- rbf.restore(inka)
predict(newrbf, newdata=iris[100,])
}
\seealso{
\link{rbf.new}
}
