% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{root.mean.squared.error}
\alias{root.mean.squared.error}
\title{Root mean squared error between two sets of numbers.}
\usage{
root.mean.squared.error(actual, estimated)
}
\arguments{
\item{actual}{True values}

\item{estimated}{Estimated values}
}
\value{
RMSE
}
\description{
Root mean squared error between two sets of numbers.
}
