% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralLayer.R
\name{sigmoid.output.function}
\alias{sigmoid.output.function}
\title{Sigmoid output function.}
\usage{
sigmoid.output.function(activation)
}
\arguments{
\item{activation}{blabla}
}
\value{
Sigmoid output value.
}
\description{
Sigmoid output function.
}
