% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralLayer.R
\name{squared.distance.activation}
\alias{squared.distance.activation}
\title{Squared Euclidean distance between weights and inputs.}
\usage{
squared.distance.activation(input, weights)
}
\arguments{
\item{input}{blabla}

\item{weights}{blabla}
}
\value{
Squared Euclidean distance
}
\description{
Squared Euclidean distance between weights and inputs.
}
