% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ss.error}
\alias{ss.error}
\title{Sum of squared errors between two sets of numbers.}
\usage{
ss.error(actual, estimated)
}
\arguments{
\item{actual}{blabla}

\item{estimated}{blabla}
}
\value{
SSE
}
\description{
Sum of squared errors between two sets of numbers.
}
