% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBF.R
\name{train.inka.formula}
\alias{train.inka.formula}
\title{Train with formula / data call}
\usage{
train.inka.formula(formula, data, spread = 0.1, normalize = TRUE, ...)
}
\arguments{
\item{formula}{blabla}

\item{data}{blabla}

\item{spread}{blabla}

\item{normalize}{blabla}

\item{...}{blabla}
}
\description{
If dependent variable is of type "factor", then it is automatically one-hot encoded.
}
