% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBF.R
\name{train.inka}
\alias{train.inka}
\title{INKA training algorithm, as in my (Kary's) PhD thesis.}
\usage{
train.inka(
  rbf,
  train.inputs,
  train.outputs,
  c = 0.01,
  max.iter = 1,
  inv.whole.set.at.end = F,
  classification.error.limit = NULL,
  rmse.limit = NULL,
  test.inputs = NULL,
  test.outputs = NULL
)
}
\arguments{
\item{rbf}{RBF network instance to use}

\item{train.inputs}{Matrix of input values of the training set}

\item{train.outputs}{Matrix of target output values of the training set}

\item{c}{"c" paremeter in INKA, i.e. minimal allowed distance between neurons}

\item{max.iter}{Maximal number of training iterations}

\item{inv.whole.set.at.end}{If TRUE, then pseudo-matrix is calculated for
entire traning set at the end.}

\item{classification.error.limit}{If specified, training ends when the
number of classification errors goes under this value. To be used for
classification tasks.}

\item{rmse.limit}{If specified, traning ends when RMSE between target
values and output values goes under (or equals) the given limit.}

\item{test.inputs}{blabla}

\item{test.outputs}{blabla}
}
\description{
INKA training algorithm, as in my (Kary's) PhD thesis.
}
