% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interfaces.R
\name{trainable.approximator.new}
\alias{trainable.approximator.new}
\title{Create "TrainableApproximator" object}
\usage{
trainable.approximator.new()
}
\value{
An object of class TrainableApproximator
}
\description{
This class is intended to
work as an object-oriented "interface", i.e. as a template for what methods
all objects of class TrainableApproximator have to implement.
}
