% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{uneven.discretizer.new}
\alias{uneven.discretizer.new}
\title{Generic discretizer "class".}
\usage{
uneven.discretizer.new(limits)
}
\arguments{
\item{limits}{blabla}
}
\value{
uneven.discretizer object
}
\description{
Returns interval indices from 1 to "nintervals"
}
