% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralLayer.R
\name{weighted.sum.activation}
\alias{weighted.sum.activation}
\title{Weighted sum activation function.}
\usage{
weighted.sum.activation(input, weights)
}
\arguments{
\item{input}{blabla}

\item{weights}{blabla}
}
\value{
Weighted sum of input, weights.
}
\description{
The most common activation function for neurons.
}
